requirements.txt
requests>=2.32.0
streamlit>=1.43.0
python-dotenv>=1.0.1
pandas>=2.2.0
